Add support to define per board custom COPR repositories
Some boards are not yet supported by the Fedora u-boot package, but should
still provide a mechanism to flash raw images on these unsupported boards.

For this reason, let's add support to define custom COPR repositories that
have firmware packages with the needed support for the given boards. Users
can then decide if want to install the COPR, and allow arm-image-installer
to install the needed firmware to boot the flashed Fedora image.

Suggested-by: Peter Robinson <pbrobinson@gmail.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>